requirements.txt
absl-py==1.1.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
cycler==0.11.0
Cython==0.29.30
easydict==1.9
faiss-cpu==1.7.2
filelock==3.7.1
flake8-import-order==0.18.1
fonttools==4.33.3
gdown==4.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
h5py==3.7.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.3
iniconfig==1.1.1
Jinja2==3.1.2
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.2
kmeans-pytorch==0.3
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.6.1
metric-learn==0.6.2
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
motmetrics==1.2.0
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
Pillow==9.1.1
pip==22.1.2
pluggy==0.13.1
prettytable==3.3.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.1.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
setuptools==63.4.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
torchfile==0.1.0
tornado==6.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==1.3.2
Werkzeug==2.1.2
wheel==0.37.1
xmltodict==0.12.0
yacs==0.1.8
zipp==3.8.0
